requirements.txt
aiohttp>=3.9.0
apify-client>=1.8.0
cryptography>=44.0.0
pandas>=2.2.0
python-dotenv>=1.0.0
requests>=2.32.0
streamlit>=1.40.0
